refactor: drop the Fluss split's unused partition name
The native log reader assigns purely by table id, partition id, and
bucket; the partition name was threaded through the split translation
only to be echoed back by a test. Also removes a leftover no-op cast.

+++ src/main/java/io/github/jordepic/streamfusion/fluss/NativeFlussSplitReader.java
@@ -129,7 +129,7 @@ final class NativeFlussSplitReader implements SplitReader<NativeSourceRecord, So
       OptionalLong partitionId = split.partitionId();
       if (partitionId.isPresent() && removedPartitions.containsKey(partitionId.getAsLong())) {
         removed.add(split);
-        buckets.add(new TableBucket(split.tableId(), partitionId.getAsLong(), (int) split.bucket()));
+        buckets.add(new TableBucket(split.tableId(), partitionId.getAsLong(), split.bucket()));
       }
     }
     for (NativeFlussLogSplit split : removed) {
